requirements/base.txt
# =========================
# Core Django
# =========================
Django==4.2.28
asgiref==3.10.0
sqlparse==0.5.3

# =========================
# Database (PostgreSQL)
# =========================
psycopg2-binary>=2.9.9

# =========================
# REST & Auth
# =========================
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==23.5

# =========================
# Channels & Redis
# =========================
channels==4.3.2
channels_redis==4.3.0
redis==5.3.1

# =========================
# Celery & Scheduling
# =========================
celery==5.6.2
django-celery-beat==2.8.1
kombu==5.6.2
amqp==5.3.1
billiard==4.2.2
vine==5.1.0
python-crontab==3.3.0
cron_descriptor==2.0.6
django-timezone-field==7.1

# =========================
# Static, Media & Files
# =========================
pillow==10.4.0
whitenoise==6.11.0
openpyxl==3.1.5

# =========================
# Security & Monitoring
# =========================
cryptography==41.0.7
PyJWT==2.10.1
sentry-sdk==1.45.1

# =========================
# Environment & Time
# =========================
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
pytz==2023.4
tzdata==2025.2
tzlocal==5.3.1

# =========================
# Admin / UI Addons
# =========================
django-jet==1.0.8
django-jet-reboot==1.3.10
django-unfold==0.7.0

# =========================
# Device / Biometric
# =========================
pyzk==0.9